Attempts
Tried to get SCV Count to a certain supply.
Also tried to fix SCV's that are idle while something under construction
to move to do something.

Tried to get more then one SCV to work on building as well.

Add Barracks as something to build.
Attempted to build marines (Couldn't progress because of issue while
working on building it does nothing else...